Refactor: Update UiText Usage and String Resources
This commit refactors the usage of `UiText` in `MoreViewModel`, `MoreScreenState`, `LocationViewModel`, and `LocationScreenState` to use the shared `UiText` class.

Additionally, it updates several composables in the `more.components` package (`LogoutItem`, `SettingItems`, `SettingItem`) to use `stringResource` for accessing string resources directly, removing the `@StringRes` annotation from parameters.

Finally, in `MoreScreen.kt`, the `SnackBar` composable's `uiText` parameter has been renamed to `text` and unnecessary `text = 0` assignments have been removed.
