+++ app/src/main/java/com/paris_2/san3a/presentation/screen/more/components/LogoutItem.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.paris_2.san3a.R
 import com.paris_2.san3a.presentation.shared.designSystem.theme.Theme
@@ -23,7 +24,7 @@ fun LogoutItem(
             onClick = {onClickItem()}
         ),
         icon = R.drawable.ic_logout_outline,
-        label = R.string.logout,
+        label = stringResource(R.string.logout),
         onClickItem = {onClickItem()},
         labelColor = Theme.colors.additional.primary.error,
         iconLabelColor = Theme.colors.additional.primary.error
