+++ app/src/main/java/com/paris_2/san3a/presentation/screen/more/moreScreen/MoreViewModel.kt
@@ -19,7 +19,7 @@ import com.paris_2.san3a.presentation.LocalAccountType
 import com.paris_2.san3a.presentation.mapper.toUserUiState
 import com.paris_2.san3a.presentation.navigation.Destinations
 import com.paris_2.san3a.presentation.shared.utils.BaseViewModel
-import com.paris_2.san3a.presentation.utill.UiText
+import com.paris_2.san3a.presentation.shared.utils.UiText
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.delay
